feat - datasource, repository 생성 및 적용

+++ app/src/main/java/com/gyurim/movieapp/presentation/data/remote/model/Movie.kt
@@ -9,5 +9,5 @@ data class Movie(
     val image: String,
     val director: String,
     val actor: String,
-    val userRating: Int
+    val userRating: Double
 )
